Reduce PHPStan baseline: 102 low-risk errors via type-accuracy fixes (#24906)
* Reduce PHPStan baseline: 66 errors across three low-risk fixes

- Intl\GenerateIntl: relax the accumulator $translations @param from
  array<string, string> to array<string, mixed> to match the mixed
  decoded-JSON values actually written, and seed it with the Intl key.
  Clears all 47 baseline entries for the file.
- Call private static/instance methods via self::/$this-> instead of
  static:: (ViewDataTable\Manager, CoreAdminHome\CustomLogo,
  CorePluginsAdmin\Controller); private methods are not polymorphic.
- Pass the '/' delimiter to preg_quote() so the pattern delimiter is
  escaped (Archiver\Request, DbHelper, Nonce, GeoIp2\GeoIP2AutoUpdater).

Removes the 14 resolved entries from the baseline.

* Reduce PHPStan baseline: 36 more errors via type-accuracy fixes

- Scheduler\Schedule\{Daily,Weekly,Hourly,Monthly,Schedule}: cast the
  date() strings passed to mktime() to int, matching the expected
  argument types (23 errors).
- Scheduler\Schedule\Weekly::setDay: widen @param to int|string, since
  callers pass day names resolved via getDayIntFromString(); the is_int()
  guard is real, not redundant.
- Application\Environment: mark $globalEnvironmentManipulator @var as
  nullable, so the six null-guards are no longer flagged always-true.
- BotTracking\AIChatbotReports::populateNumerics: widen the by-ref
  $tables @param to array<string, int|DataTable>, since it writes int
  metric values into the accumulator (7 errors).

Removes the 25 resolved entries from the baseline.

* Fix stale PHPStan inline-ignore in CoreUpdater\Updater

The // @phpstan-ignore notIdentical.alwaysFalse on the
PIWIK_INCLUDE_PATH !== PIWIK_DOCUMENT_ROOT check became stale under the
newer PHPStan: the two path constants are aliased in the analysis
bootstrap, so whether the comparison folds to always-false depends on
the analyzer version, leaving the inline ignore unmatched on CI.

Read both constants into string-typed locals so the comparison is
treated as a genuine runtime string check in every environment, and drop
the no-longer-needed inline ignore.

+++ core/Scheduler/Schedule/Monthly.php
@@ -65,12 +65,12 @@ class Monthly extends Schedule
 
         // Adds one month
         $rescheduledTime = mktime(
-            date('H', $currentTime),
-            date('i', $currentTime),
-            date('s', $currentTime),
-            date('n', $currentTime) + 1,
+            (int) date('H', $currentTime),
+            (int) date('i', $currentTime),
+            (int) date('s', $currentTime),
+            (int) date('n', $currentTime) + 1,
             1,
-            date('Y', $currentTime)
+            (int) date('Y', $currentTime)
         );
 
         $nextMonthLength = date('t', $rescheduledTime);
